Fix importing photos for all contacts
Django can already select all matching rows. So use that instead of
implementing a custom selection for "all" contacts.

+++ core/models.py
@@ -689,7 +689,7 @@ class ImportContactPhotosTask(TaskRQ):
     LOG_TO_FILE = False
 
     contact_ids = ArrayField(
-        base_field=models.IntegerField(),
+        base_field=models.BigIntegerField(),
         null=True,
         blank=True,
         help_text=(
